Add flavors for Huawei and GMS
* Set enviroment for Huawei devices testing
