+++ Examples/OneSignalDemo/app/src/huawei/AndroidManifest.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:amazon="http://schemas.amazon.com/apk/res/android"
+    package="com.onesignal.sdktest">
+
+    <application
+        android:name=".application.MainApplication">
+
+        <service
+          android:name="com.onesignal.sdktest.notification.HmsMessageServiceAppLevel"
+          android:exported="false">
+          <intent-filter>
+            <action android:name="com.huawei.push.action.MESSAGING_EVENT" />
+          </intent-filter>
+        </service>
+
+    </application>
+
+</manifest>
